moved cert generation file inside fixture folder
