+++ tests/ginkgo/fixture/certificate/fixture.go
